Changed price type from float to decimal

+++ Database Project/Entities/Detail.cs
@@ -6,5 +6,5 @@ using System.Threading.Tasks;
 
 namespace Database_Project.Entities
 {
-    public record Detail(int DetailId, float Price, int Quantity, int OffertId, int SellId);
+    public record Detail(int DetailId, decimal Price, int Quantity, int OffertId, int SellId);
 }
